test: extract TestDataProvider
Extract TestDataProvider to separate class
Add test input for CompleJob
