+++ Client.UnitTests/SendWithRetryParameterizedTest.cs
@@ -12,22 +12,6 @@ using Zeebe.Client.Api.Responses;
 
 namespace Zeebe.Client
 {
-    public delegate IFinalCommandWithRetryStep<T> RequestCreator<T>(IZeebeClient zeebeClient);
-
-    public class TestDataProvider
-    {
-        public static IEnumerable<TestCaseData> Provider()
-        {
-            yield return new TestCaseData(
-                new CancelProcessInstanceRequest
-            {
-                ProcessInstanceKey = 12113
-            }, new CancelProcessInstanceResponse(),
-                (RequestCreator<ICancelProcessInstanceResponse>)
-                (zeebeClient => zeebeClient.NewCancelInstanceCommand(12113)));
-        }
-    }
-
     [TestFixture]
     public class SendWithRetryParameterizedTest : BaseZeebeTest
     {
